Clase 18: pwa & deploy

+++ ToDo-AppV1.2/scripts/taks.js
@@ -8,6 +8,9 @@ if (!localStorage.jwt) {
 
 /* ------ comienzan las funcionalidades una vez que carga el documento ------ */
 window.addEventListener('load', function () {
+  /* ---------------- Inicializamos la libreria AOS ---------------- */
+  AOS.init();
+  
   /* ---------------- variables globales y llamado a funciones ---------------- */
   const btnCerrarSesion = document.querySelector("#closeApp")
   const formCrearTarea = document.querySelector(".nueva-tarea")
@@ -30,13 +33,34 @@ window.addEventListener('load', function () {
   /* -------------------------------------------------------------------------- */
 
   btnCerrarSesion.addEventListener('click', function () {
-    const cerrarSesion = confirm("¿Estás seguro de que deseas cerrar sesión?")
+    // const cerrarSesion = confirm("¿Estás seguro de que deseas cerrar sesión?")
 
-    if (cerrarSesion) {
-      // limpiar el LocalStorage y redireccionar al login
-      localStorage.clear()
-      location.replace("./index.html")
-    }
+    // if (cerrarSesion) {
+    //   // limpiar el LocalStorage y redireccionar al login
+    //   localStorage.clear()
+    //   location.replace("./index.html")
+    // }
+
+    Swal.fire({
+      title: "¿Estás seguro de que deseas cerrar sesión?",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "¡Si, Confirmo!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          '¡Hasta luego!',
+          'Te esperamos pronto.',
+          'success'
+          )
+          setTimeout(() => {
+            localStorage.clear()
+            location.replace("./index.html")
+          }, 2000);
+      }
+    });
   });
 
   /* -------------------------------------------------------------------------- */
@@ -161,7 +185,7 @@ window.addEventListener('load', function () {
         
         /// renderizo las tarjeas completadas
         tareasTerminadas.innerHTML += `
-          <li class="tarea">
+          <li class="tarea" data-aos="flip-up" data-aos-duration="1500" >
             <div class="hecha">
               <i class="fa-regular fa-circle-check"></i>
             </div>
@@ -178,7 +202,7 @@ window.addEventListener('load', function () {
       } else {
         // renderizo la tarjetas pendientes
         tareasPendientes.innerHTML += `
-          <li class="tarea">
+          <li class="tarea" data-aos="fade-right" data-aos-duration="2500" >
             <button class="change" id="${tarea.id}"><i class="fa-regular fa-circle"></i></button>
             <div class="descripcion">
               <p class="nombre">${tarea.description}</p>
@@ -252,22 +276,38 @@ window.addEventListener('load', function () {
     btnBorrarTarea.forEach(boton => {
       //a cada boton de borrado le asignamos la funcionalidad
       boton.addEventListener('click', function (event) {
-        const id = event.target.id;
-        const url = `${urlTareas}/${id}`
+        Swal.fire({
+          title: "¿Confirma que desesa eliminar tarea?",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "¡Confirmo!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            const id = event.target.id;
+            const url = `${urlTareas}/${id}`
+    
+            const settingsCambio = {
+              method: 'DELETE',
+              headers: {
+                "Authorization": token,
+              }
+            }
+            fetch(url, settingsCambio)
+              .then(response => {
+                console.log("Borrando tarea...");
+                console.log(response.status);
+                //vuelvo a consultar las tareas actualizadas y pintarlas nuevamente en pantalla
+                consultarTareas();
+              })         
 
-        const settingsCambio = {
-          method: 'DELETE',
-          headers: {
-            "Authorization": token,
+              Swal.fire(
+                 "¡Tarea eliminada!"
+              );
           }
-        }
-        fetch(url, settingsCambio)
-          .then(response => {
-            console.log("Borrando tarea...");
-            console.log(response.status);
-            //vuelvo a consultar las tareas actualizadas y pintarlas nuevamente en pantalla
-            consultarTareas();
-          })
+        });
+
       })
     });
   }
